+++ src/ui/PlayerBar.cpp
@@ -538,6 +538,10 @@ PlayerBar::PlayerBar(QWidget *parent)
     QAction *libraryShuffleSettings = playbackMenu->addAction(QStringLiteral("Library shuffle..."));
     QAction *radioShuffleSettings = playbackMenu->addAction(QStringLiteral("Radio shuffle percent..."));
 
+    auto *mixesMenu = new QMenu(QStringLiteral("Mixes"), this);
+    QAction *rediscoveryMix = mixesMenu->addAction(QStringLiteral("Play Rediscovery mix"));
+    QAction *deepCutsMix = mixesMenu->addAction(QStringLiteral("Play Deep cuts mix"));
+
     auto *mpdMenu = new QMenu(QStringLiteral("MPD"), this);
     QAction *mpdSource = mpdMenu->addAction(QStringLiteral("Configure MPD source..."));
     QAction *mpdImport = mpdMenu->addAction(QStringLiteral("Import MPD library metadata"));
@@ -639,7 +643,7 @@ PlayerBar::PlayerBar(QWidget *parent)
     m_alwaysShowTray = settingsMenu->addAction(QStringLiteral("Always show system tray icon"));
     m_alwaysShowTray->setCheckable(true);
 
-    const QVector<QMenu *> styledMenus{compactMenu, fileMenu, ratingTagsMenu, scanPowerMenu, viewMenu, queueMenu, playlistMenu, playbackMenu, mpdMenu, historyMenu, scrobblersMenu, settingsMenu};
+    const QVector<QMenu *> styledMenus{compactMenu, fileMenu, ratingTagsMenu, scanPowerMenu, viewMenu, queueMenu, playlistMenu, playbackMenu, mixesMenu, mpdMenu, historyMenu, scrobblersMenu, settingsMenu};
     for (QMenu *menu : styledMenus) {
         styleMenu(menu);
     }
@@ -649,6 +653,7 @@ PlayerBar::PlayerBar(QWidget *parent)
     compactMenu->addMenu(queueMenu);
     compactMenu->addMenu(playlistMenu);
     compactMenu->addMenu(playbackMenu);
+    compactMenu->addMenu(mixesMenu);
     compactMenu->addMenu(historyMenu);
     compactMenu->addMenu(settingsMenu);
 
@@ -679,6 +684,7 @@ PlayerBar::PlayerBar(QWidget *parent)
     m_menuBar->addMenu(queueMenu);
     m_menuBar->addMenu(playlistMenu);
     m_menuBar->addMenu(playbackMenu);
+    m_menuBar->addMenu(mixesMenu);
     m_menuBar->addMenu(historyMenu);
     m_menuBar->addMenu(settingsMenu);
     menuStripLayout->addWidget(m_menuButton);
@@ -839,6 +845,8 @@ PlayerBar::PlayerBar(QWidget *parent)
     connect(playbackResume, &QAction::triggered, this, &PlayerBar::playbackResumeRequested);
     connect(libraryShuffleSettings, &QAction::triggered, this, &PlayerBar::libraryShuffleSettingsRequested);
     connect(radioShuffleSettings, &QAction::triggered, this, &PlayerBar::radioShuffleSettingsRequested);
+    connect(rediscoveryMix, &QAction::triggered, this, &PlayerBar::rediscoveryMixRequested);
+    connect(deepCutsMix, &QAction::triggered, this, &PlayerBar::deepCutsMixRequested);
     connect(linkRoots, &QAction::triggered, this, &PlayerBar::linkRootsRequested);
     connect(mpdSource, &QAction::triggered, this, &PlayerBar::mpdSourceRequested);
     connect(mpdImport, &QAction::triggered, this, &PlayerBar::mpdImportRequested);